battlebots/requirements.txt
Flask>=0.10.1
Markdown>=2.6.5
Flask-WTF>=0.12
Flask-SQLAlchemy>=2.1
SQLAlchemy-Migrate>=0.10
Flask-Login>=0.3.2
